# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DocSecur API",
        "description": "API REST pour la plateforme de gestion médicale DocSecur. Authentification via Bearer Token (Laravel Passport).",
        "contact": {
            "email": "support@docsecur.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/medecin/demandes-analyse": {
            "get": {
                "tags": [
                    "Médecin - Analyses"
                ],
                "summary": "Lister les demandes d'analyse du médecin",
                "description": "Retourne la liste paginée des demandes d'analyse émises par le médecin connecté.",
                "operationId": "d245d1afbeaafe22cc2ebb9188f668b5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des demandes paginée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Médecin - Analyses"
                ],
                "summary": "Créer une demande d'analyse",
                "description": "Envoie une demande d'analyse à un laboratoire pour un patient.",
                "operationId": "87d0d6c828f0159bf6bab0840a42cc79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "laboratoire_id",
                                    "type_analyse"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "laboratoire_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type_analyse": {
                                        "type": "string",
                                        "example": "Numération Formule Sanguine (NFS)"
                                    },
                                    "urgence": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Patient fébrile depuis 3 jours"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Demande envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Demande d'analyse envoyée"
                                        },
                                        "demande": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medecin/demandes-analyse/{id}": {
            "get": {
                "tags": [
                    "Médecin - Analyses"
                ],
                "summary": "Détails d'une demande d'analyse",
                "operationId": "ceec71b5c73ee80037e3c687a7009ab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la demande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Médecin - Analyses"
                ],
                "summary": "Supprimer une demande d'analyse",
                "operationId": "10114c823f522dd299d4d5535cadf2ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Demande supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Demande supprimée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}